requirements.txt
alembic==1.18.1
azure-ai-contentsafety==1.0.0
azure-core==1.38.0
azure-identity>=1.25.3
azure-keyvault==4.2.0
azure-keyvault-certificates==4.10.0
azure-keyvault-keys==4.11.0
azure-keyvault-secrets==4.10.0
black==26.1.0
fastapi==0.128.0
openai>=2.28.0
pydantic==2.12.5
pydantic-settings==2.12.0
pyflakes==3.4.0
pyodbc==5.3.0
python-dotenv==1.2.1
SQLAlchemy==2.0.45
starlette==0.50.0
tenacity==9.1.2
uvicorn==0.40.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
azure-search-documents==11.6.0
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1